backend/requirements.txt
alembic==1.10.2
amqp==2.6.1
aniso8601==9.0.1
asgiref==3.5.2
attrs==22.2.0
Babel==2.11.0
bcrypt==4.0.1
billiard==3.6.4.0
black==23.3.0
pydantic
uvicorn
fastapi
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==4.6.11
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.1
pathspec==0.11.1
platformdirs==3.5.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
SQLAlchemy-serializer==1.4.1
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
vine==1.3.0
Werkzeug==2.2.3
